Refactor code to make it more readable.
